Minor Changes
Changed location of the player, added a tag for the floor to help with
navigation later, and messed around with ideas and code to remove the
press E text. I know how to do it, but I will be downtown tonight. If I
get a minute, I know how to do it.