Turned off light by contact with enemy
I had to add a sphere collider to the light to let the enemy find the
Light tag. I also increased the enemy stopping distance from 2 to 2.1 to
prevent him from running into the player. I increased the radius of the
sphere collider on the light from .2 which looked nice to .5 to match
this new enemy distance. This is temporary and once we figure out the
sky rocket bug then we can change those parameters.