Began work on destroying/disabling pickup item
Entered in the bool function that requires text to be shown and e to be
pressed for the bool to work. Cannot figure out how to set
CollectingItem.gameObject.SetActive(false); to the game object. I said
CollectingItem was a game object, but I cannot get the sphere we need
gone to be that game object.