Fixed enemy light collider and remove collectable
Accidentally realized I was very close at solving the pressE to remove
object, and at the same time, figured that out by figuring out how to
turn off the flashlight with the player collider and removed the light
collider. Fixed two issues at the same time. Awesomesauce. This should
be my last change to the Master. I go to work here soon so if you wanted
to create the work around branch, that would be sweet