Substantial updates to enemy
I re-baked the nav mesh so now the enemy avoids all obstacles. When I
was messing with the environment before, I was sad to see it run into a
wall and stop chasing the player. So now it can smartly figure out the
best path. It cannot jump down however because that is only available in
pro. I created a boolean to make it so it changes navigation when near
the player. And I also increased the speed of the enemy to 10 when it
changes its navigation path. Probably later tonight I will set up the
reverse for when the player touches the enemy to get the glasses back. I
am also very forgetful about making a new branch and I keep just making
edits straight to master. Sorry about that. I will work on it.