Attempted returning light fix
Ran into a glitch with the flare and collider. Also realized flaw in my
original plan because I am getting the light back as soon as the enemy
takes it. I need to either add a delay to the time it takes to return
light, or figure out NavMeshPathStatus.PathComplete so this cannot be
activated until the enemy gets to its path.