Closer
Light turns off when player and enemy meets. Light turns back on when
player and enemy meet again and enemy waits (currently) 3 seconds before
chasing the player again. The total time will be increased to around 15
or 20 seconds. The problem is getting the whole thing to start over
again without duplicating code. Need to find a way to recall the section
void OnTriggerEnter.