Large update including addition of night vision
New script and light to create night vision. Currently I have it set up
that if you press g you switch from flashlight to night vision and back
again. Night vision has the grainy look, but non moving. Once I upgrade
to Unity pro, the static component will move. I may want to darken the
sides of the night vision because if you think about glasses, they do
not cover the eyes fully like goggles do. So there would be some space
in the sides that show up dark. Later I want to make it so you press a
different button to turn on flashlight and night vision, but one cannot
be on while the other is on. That is just an if statement. Eventually,
when the other spectrums are made, I want all of the spectrums to be
assigned by holding down the middle mouse button and mousing over which
spectrum you want. Updates to come for night vision still.