Enemy disappear/dust&dirty lens/g key/toggleflash
A dust texture is on in the ToggleFlashlight script which has been moved
to the player instead of on the blue light itself. Updated the coding so
pressing g switches between light and night vision. q still controls
blue light. The enemy disapears when in the night vision mode. But you
can still run into it, and it is functional. Does not remove night
vision yet. Dust texture is set to 0 alpha for the time being. Looks
good over blue light on a .35 alpha setting, but because it has a black
background, it causes things to go dim. Looking to change that. I bought
an asset for a dirty lens flare, which gives a really good glass and
dirt look when looking at a light. Not sure when to implement, but it
was cheap.