Massive update to inventory
I followed a tutorial series for the inventory, but apparently the
creator put out the last video in the series over 8 months ago and said
he would return but hasn't... So there is a lot I need to do to get this
perfect. In this update is pressing tab to open the inventory. Temporary
icons for the items. A way to drag and rearrange the items. Item
classifications. The ability to double click. Double clicking the key
lets you "use" the item. A save and load mechanic has been added to the
inventory. Slots and boxes were added as well as Item id's. Collect item
was edited and prepared to be merged with AddItem();