Closer to adding item
The Collect Item script located in the first person controller game
object has a slot for inventory. I have put the inventory in that slot,
but after hitting play it takes it away. If I put the inventory back in
that slot after hitting play, it works. Need to figure out why it is
deleting it.