I r smrt
I follow a tutorial on how to make an inventory system. Of course the
whole project is just that so he creates an empty game object and then
adds the scripts to it. Which is what I did, and then tried to have a
script call that game object and that script. Why don't I just put that
script on the player like it should be instead of putting it on a
different game object all together? Dumb. I reordered the script
sequence so the inventory script works first.