Rotation Hell
Had to do a lot of research to find a way to get a door to open in the
negative degrees. It was always turning to 270 when I wanted -90. Then I
had to do more research in order to find how to change the pivot
location. It was a nightmare. In the research I found out a way to have
an object follow the player though. So I threw that script onto "Enemy2"
for the time being. I had to duplicate the OpenDoor script and change it
to OpenDoorL in order to get the doors to open differently. This
probably needs to be optimized, as well as the box colliders. Lots of
work needs to be done on this still, but it was a full day of research
just to get it to rotate backwards.