Condensed scripts/fixed loophole/added button
Moved AlwaysWatching script. Moved all the different content in
OpenDoorL into the OpenDoor script. That way I am not duplicating
scripts. Deleted OpenDoorL script. Added a button to be the trigger
instead of the door. Added a boolean that makes it so if I am not near
the collider, pressing E doesn't open the doors again. need to be in the
collider only to open the door.