More pickup adjustments - pickup/return key
Big thing is now you cannot pick up items when you are standing on an
item and you are in the inventory. You need to be in the game mode in
order to pick that object up. Made a way where if you press E on
anything that isnt the door, the key returns to your inventory. Then
fixed it so if you are standing on an item, you just pick up the item
but the key does not get returned. Key also leaves the players hand when
it is returned.