Allowing one item at a time #1
Started the Use Gun method. However when selecting the gun when you have
the key out, you get both out and it looks doofy. I have been trying to
get one to go away and return to the inventory when the other is out,
but it has been a challenge. I am wondering if I shouldn't even bother
and just let both of them out? Left hand key and right hand gun. But
then again I will run into the issue later when I try to use a third
item. So I need to fix it anyway.