Changed Tags, Moved code, Swap items, Fixed return
Added more tags to make things more specific. The door was looking for
the key on the ground and not the key the the player was holding. Made
haveKey = false when not near the door. This allows the player to put
away the key when they press E not near the door. Pressing E near the
door when holding the key no longer returns the key. Having either the
key or the gun out and then switching to the other item returns the
previous item to the inventory. Fixed all instances of when that wasn't
the case. Moved all of that excess code and condensed it by moving it
into the inventory class. Simple is best.