Added dust and motion to the first door
Particle systems have .001 and -.001 gravity to give floating effect.
Will tweak for part of the graphics work. Just wanted to give it a
start. Got one of the doors to start moving. Adding all of the elevator
doors scripts to the elevator itself. Will have to add more to it.