The Devastating Update
Where to begin on this one? Unity updated. This caused every game
object, value, and component that required materials to be dragged to
it, like a public variable, to get removed. There were no wait times,
not objects attached to the reference, and so on. I had to go through
every script and object to apply a way to prevent that from happening
again. I will explain the main changes here.
Sight: Turned this off because I wasn't ready to start Raycast until I
fixed everything
Bullet: Didn't automatically set up the bullet prefab because it is in
the assets and not in the hierarchy. Changed shooting power to be 150 at
default.
CollectItem: The items that can be picked up should find their items tag
and then assign that way instead of dragging it to it.
Elevator: This one sucked because I didn't know all of these numbers by
heart. I set the elevators transform to equal where it currently is.
Found the down and up positions and locked those in. Will have to do the
same with the other floors. Open and closed position was still on my
board fortunately.
EnemyDamage: Thought about grabbing the light from a different class and
changed my mind. Now set it so the lights find their tag. Found out they
can not do this if the game object is false in the beginning. So this
caused me to turn on the rest of the objects. They turn off once the
game starts. Awake function and all that. Added the wait times back to
the enemy.
FlickeringLight: Fortunately this also taught me how to find objects
based on their name and how to grab their components.
Inventory: Came back to this one many times. Slots are 5 and 2. Didn't
set them because I wasn't sure at the time. I am going to go back and
set those right after this. Find the holding objects. Set tapspeed
instead of just typing it in.  Oh, and I found out an issue I didn't
know about. When you make Time.time = 0, that stops the time altogether
so double tapping was useless because it happened at the same 0 time.
Changed it to realtimeSinceStartup.
OpenDoor: Did not set doors transforms! I just threw them in. I could
use the getComponent though.
TestingNightVision: Texture texture is very vague. Changed it so I know
what is supposed to be dragged in there. Do not have this automatically
set. No texture is besides the icons in inventory. Renamed enemy to the
actual enemy it is supposed to be. Helps in long run. Gave the lights
tags so they could be found.
ToggleFlashlight: Hahaha. This is how I know I have improved in coding.
I was using Light and LensFlare instead of gameObjects. So I was calling
each item twice just to use their other components. Plus this was all
named out of wack. Lightintensity1 is the flashlight? How would anyone
else know that? Changed those names to make sense. This whole script
looked like a disaster. Changed the name of the dust, organized the
numbers, set the intensities.
SpectrumTesting: Everything got default set to 0. Added plenty of tags.