Major update #1 to add items
Added the holding items. Created booleans for active items and swapped
items. Set holding items to be false. Added explanations to inventory
menu functions. Started but backtracked on combining items. Instead of
doing use key and use gun and use all other items, I shortened it down
to use item and cant use item. Added the cases for the items that can be
used and cant be used. Makes them active and makes the object visible.
Haven't made them swap yet. Added combine item type to the Item class.
Updated description, added combined items, and fixed item types to the
item database. Moved the gameobjects into place and attached to the
camera. Removed tags for the items that were going to be held, but
didn't need to be because they are going to get combined. Updated other
tags.