Major update #1 to ItemEffects
Moved all pickup items to be on the player when the game starts. Makes
it easier to pick them up in one button instead of walking around to
pick them up. Made changes to the bullet prefab including changing it to
dynamic continuous so the bullets don't fall through the floor. Made
changes for the bullet in the bullet script as well as added the
throwing components for the rock. Changed the location for the thrown
rock and gun. The gun now follows the characters head instead of the
stationary point near the chest. Rock throws weaker than the bullet.
Made the rock dynamic continuous as well so they don't fall through the
floor. The rock stays on the field, but has odd rolling mechanics.
Doesn't roll down hill correctly and keeps rolling on the ground without
stopping. Need to make the rock have more friction and unique collider
parameters. Made fixes that prevent the gun from being shot while in the
inventory, as well as when you switch out of the inventory. Added the
mother to the top of the ramp. When the little girl gets near the mom,
it stops going after the player and then goes away. Added
PlasticFeatures script to make it so handing the anatomy item to the
plastic model makes him die.