Quick change
Realized I was testing find anatomy in the update parameter instead of
the start one. My reasoning behind this was it wasn't recognizing the
anatomy idea, and I believed it was because it is set to false in the
beginning. This isn't the case though, so I changed it back.