Surgeon and Lights
Added death sequence to the surgeon. Began work on getting the
flickering lights to go static upon pausing the game. This effect was
achieved, but I want it to happen with a random number generator and
varying brightness. Could this be done with the way the flashlight kind
of worked? Something like setting a variable to be the intensity it is
at by the time they hit tab?