Elevator/Key, Flickering lights, Game controller
Added Elevator key item to all components of collect item and inventory
giving it its own function as well as interaction with others. Lots of
attempts have been made to fix the elevator. I added the script player
controller to act as a puppet master for colliders and game mechanics.
This separates the collider inside the elevator and the one in-front of
the elevator. I figured out how to make a random number generator in
order to get a more realistic lighting effect when pausing the game and
looking at the flashing light.