Animation and Elevators
Created animation to use on the doors instead of positioning and
vectors. This makes life easier. I removed everything from the elevator
script because it was starting to get complicated by having some
components be in player controller, and some in elevator. So everything
is merged into player controller now. Removed a lot of unnecessary and
dead code. I plan on going back tomorrow to top off all of the elevator
components. Like make it one button, only have the elevator activate
when the door is closed, and so on.