Touch-ups and Surgeon moves
Made changes to the elevator so the door cannot be activated multiple
times in a row. Can only activate when the animation is not running.
Changed the t button to 1 indicating first floor.  Fixed old light
intensity in TestingNightVision. Added navmesh agent and searching
ability to the Surgeon. He now runs from limb to limb and speeds up each
time. Fine tuning will be later.