Major Raycast update
Made bullethole prefab and script to remove it after a couple of
minutes. Only works on objects tagged Environment. Changed the mass of
the bullet to be non-existent so it isn't pushing enemies around.
Created separate rock script away from the bullet script. Each prefab
needed to be with its object, not together. Moved bullet and rock
scripts. Added script to bullet that creates bulletholes if walls are
shot, or makes the oldManShot bool true if the old man is shot. Added
future code for hitting objects that could be pushed around. Temporarily
removed hospital girls speed. She was getting in the way. Removed enemy
movement script. Was not needed. Looking directly at the old man gives
you the oldManSeen bool. Having him within the angle gives you the
peripheral bool. Looking past the oldManViewAngle or looking up or down
too far removes that bool.