Better Update
OldMan script is gone and has been replaced with the Wander script. The
script makes it so the old man wonders from point to point in the
beginning, determined by random number. Then if he is spotted, he speeds
up and faces the direction of the player. Reaching the player gives him
his target angle, and the spots that he was wondering around in scatter
to the new spawn locations and he sprints there based on the angle he
was heading towards. Need to figure out why when the old man runs away
it looks like he is drunk...