Timed Teleport
Now every 1.5 seconds the jumper will check to see if it is more than 2
away and then jump the calculated distance towards the player. This is
done by starting a coroutine at the beginning of the game and having it
go through an if statement before getting sent through the coroutine
again. More updates to come. I am thinking about making the time until
next jump a randomized variable.