Jumper/HG/CharacterMotor Major Update
Removed Java Script scripts in order to bring in C# versions of them.
This was necessary to access the motor script of the player. The
standard scripts were moved into the scripts folder. CharacterMotorC and
FPSInputControllerC are the new version and were attached to the player.
The controls were different so I lowered the speed, raised gravity, and
lowered jump boost. Still not the same as it was since the plastic
models head now throws me off when it didnt do that before. Made some
smarter changes to EnemyDamage. Twice I was calling for the player, so I
made Main be one thing like position and player be the script I needed.
Temporarily changed the HG talking speed. Made the code that makes it so
when the HG catches the player and the player is looking at her, she
teleports behind the player and shows up again after turning 90 degrees
either direction. More on her later. Condensed Jumper script a lot.
Removed Back and Right controls. Corrected for when the player is at a
different angle on the Z plane so x and z are interchangable. Being far
away from the jumper makes it go into fast mode.  Changed the values of
the ranges. Made TeleportPlayer() in order to have the player be unable
to move when the jumper catches them. Added hGInsight to
playerController. Added new game objects to the sides and back of the
player.