Fixed Echo. Thrown Rock. Organization. Progress
Moved scripts to the script folder. Overhauled the EchoSphere script so
it works. Could not get EchoSpheres and EchoSphere2 to work. I will find
out my own way to get multiple spheres. Added a new script to the thrown
rock. The thrown rock can now be picked up like a regular item and
thrown again. Changed components to make 1,2,3 be the triggers for the
lights, and organized a bit so code wasn't being used multiple time.
Learned a ton. Made arrays work in order to find all of the environment
options, grabbing their renderer, and then changing it to the echo
format. Created a ping location so the echo doesn't follow the player.
Will apply this to the rock as well when I get multiple echos. Switching
to sonar vision turns off all other lights. Created to-do list for the
rest. Much better day today than yesterday.