Started return materials
The materials need to go back to what they were originally when
switching out of sonar mode. I currently have them all going back to
default, but the floor wasn't default. There may be a way to store what
was there at the start of the game and then just return  what was there.