Rock Physics, FloorMaterial, Sonar False, Elevator
Made it so when you are in sonar mode and you switch out, the material
changes back. The floor had to have its own material because it is
different than the rest. Need to figure out a way to store all values,
or else I will have to set all values manually. Changed the rock prefab
from a round collider to a square collider with a physics component that
gives it some bounce and stops moving after time. Fixed the elevator
colliders so walking into the elevator lets you send it up instead of
whenever.