First attempt at multiEcho
What I have done is separated and merged a lot of code into two scripts
and two shaders. One of the scripts and shaders do what they did before
with the single echo. This was kept just to make sure I didn't mess it
up. I have a second script that has the elements for multi echos. Just
need to get it to work though. Currently it shows up, but at origin, and
not multi. Need to figure out lists better as well as adding to i. I
organized a couple scripts, but that has been set to the side for now.