OpenDoor and Elevator moved into Environment Script
Made changes to organization of the environment items that were causing
issues, such as the elevator doors position moving to a different
location. Merged the OpenDoor and Elevator components into the
environment script. Changed environment script to environmental. Made
small changes to echospherez to return the floor and return default
materials. All OnTrigger functions need to stay on the player. The
Activation switch triggers was moved to the player so it activates when
it hits the switch instead of when the switch finds the key. Because
StartCoroutine of WantToOpenDoor() is not single function required, I
made it so update can trigger it when the player brings out the key and
they were already standing at the door.